Adjust scores on Cape sigs
On the injection sigs it may be more useful to have scores that are higher. While it is nice to group them all together to indicate they are CAPE related it may be better to increase score in case they are the only signatures firing so the score reflects it is malicious.