Rewrite of Powershell command signature
This signature as it stands in its current form was not working. Took matches from it and now instead of a lit of anomalies just dumps the powershell command (for now) which I think is more useful.

Also implemented a base64 decoding function which while not perfect should be able to extract and present decoded base64 strings although this requires it is following a base64 string so for instance "-enc BASE64" not a variable like "$content = BASE64 -enc $content". A solution to this would be extract and decode all base64 strings but with evasion, dosfuscation etc. this may be difficult. As such between this signature and some of the command line signatures most suspicious powershell commands are detected IF they are using these features or obfuscation.

I will improve this signature over time. I did consider attempting my first ML project with this after being inspired here https://www.fireeye.com/blog/threat-research/2018/07/malicious-powershell-detection-via-machine-learning.html but I think this is not entirely necessary the now as enough detections exist and also I don't have enough malicious and varied examples of malicious powershell commands let alone real world benign examples for training and edge cases.